requirements.txt
absl-py==1.3.0
aicore==0.0.3
alembic==1.14.0
annotated-types==0.7.0
anyio==4.9.0
appnope==0.1.3
asttokens==2.0.8
astunparse==1.6.3
async-generator==1.10
attrs==22.1.0
awscli==1.27.94
backcall==0.2.0
beautifulsoup4==4.11.1
blinker==1.5
boto3==1.26.94
botocore==1.29.94
branca==0.7.0
Brotli==1.1.0
cachetools==5.2.0
certifi==2024.8.30
cffi==1.17.1
chardet==4.0.0
charset-normalizer==2.1.1
colorama==0.4.4
contourpy==1.0.6
cryptography==45.0.4
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
distro==1.8.0
docopt==0.6.2
docutils==0.16
entrypoints==0.4
exceptiongroup==1.3.0
executing==1.1.1
flatbuffers==22.9.24
folium==0.15.1
fonttools==4.38.0
gast==0.4.0
google-auth==2.13.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==2.0.2
grpcio==1.50.0
h11==0.16.0
h2==4.2.0
h5py==3.7.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==2.10
ipykernel==6.16.0
ipython==8.5.0
jedi==0.18.1
Jinja2==3.1.2
jiter==0.10.0
jmespath==1.0.1
joblib==1.2.0
jupyter-core==4.11.1
jupyter_client==7.4.2
kaitaistruct==0.10
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==14.0.6
Mako==1.3.6
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mplsoccer==1.4.0
nest-asyncio==1.5.6
numpy==1.23.4
oauthlib==3.2.2
openai==1.86.0
opencv-python==4.6.0.66
opt-einsum==3.3.0
outcome==1.2.0
packaging==21.3
pandas==1.5.2
parso==0.8.3
patsy==1.0.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
pipreqs==0.4.11
potracer==0.0.4
prompt-toolkit==3.0.31
protobuf==3.19.6
psutil==5.9.2
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
Pygments==2.13.0
pyOpenSSL==25.1.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
PyYAML==5.4.1
pyzmq==24.0.1
requests==2.25.1
requests-oauthlib==1.3.1
rsa==4.7.2
s3transfer==0.6.0
scikit-learn==1.2.0
scipy==1.9.3
seaborn==0.13.2
selenium==4.6.0
selenium-wire==5.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==2.0.5.post1
stack-data==0.5.1
statsmodels==0.14.4
tabula-py==2.7.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==2.0.1
threadpoolctl==3.1.0
torch==1.13.1
torch-tb-profiler==0.4.1
tornado==6.2
tqdm==4.64.1
traitlets==5.4.0
trio==0.22.0
trio-websocket==0.9.2
typing-inspection==0.4.1
typing_extensions==4.14.0
understatapi==0.6.1
urllib3==1.26.5
wcwidth==0.2.5
webdriver-manager==3.8.5
Werkzeug==2.2.2
wrapt==1.14.1
wsproto==1.2.0
xyzservices==2023.10.1
yarg==0.1.9
zstandard==0.23.0
